### 2025-02-19 #7796 @editorialbot
[PRE REVIEW]: GCol: A High-Performance Python Library for Graph Colouring
**Submitting author:** @Rhyd-Lewis (<a href="http://orcid.org/0000-0003-1046-811X">Rhyd Lewis</a>)
**Repository:** https://github.com/Rhyd-Lewis/GCol
**Branch with paper.md** (empty if default branch): josspaper
**Version:** v 1.4.2
