### 2025-02-20 #7796 @Rhyd-Lewis
Hi @danielskatz - thanks for your message. I have been researching graph colouring for about 20 years and wrote a book about it in 2015 (and a [second edition](https://link.springer.com/book/10.1007/978-3-030-81054-2) in 2021). As part of this, I maintain a c++ library of graph coloring code to accompany the book.

I decided to start working on a python library in the summer of 2024. At first, I thought I'd just add some stuff to NetworkX, but soon I realised I'd be able to do a lot more by writing a new library built on top of NetworkX. The GCol library also extends the C++ code by allowing face coloring, edge colouring, visualisations etc, as you'll see in the [documentation](https://gcol.readthedocs.io/en/latest/).

Development was done offline, so I only created the repo fairly recently.

Thanks

Rhyd
