### 2025-02-20 #7796 @danielskatz
Thanks @Rhyd-Lewis. I'm going to ask other JOSS editors to discuss if this is clearly of sufficient scope for JOSS to review it. You should hear back in a week or two.
