### 2025-02-22 #7796 @sashakolpakov
Does the Editor understand that given the domain significance (graph colouring is a classical domain) and author's profile (a world-renowned expert), a scope request lacks all seriousness? Forgive my straightforward comment, but it is exactly the right time for one.
