### 2025-02-22 #7796 @arfon
@sashakolpakov – please refrain from interfering with the JOSS editorial process. I consider this interjection a violation of JOSS' [code of conduct](https://github.com/openjournals/joss/blob/main/CODE_OF_CONDUCT.md). The editorial team is responsible for ensuring that the submissions we review here are in scope for our journal. This has nothing to do with the background, history, renown of the submitting author.

@Rhyd-Lewis – as @danielskatz mentioned above, the scope review is a standard part of our processes and should complete relatively soon. Thanks in advance for your patience.
