### 2025-02-22 #7796 @sashakolpakov
@arfon Thank you for your quick response, and for keeping the process transparent for the community. Having your actions questioned is a part of it. Indeed, the submission appears academically solid to me and, moreover, useful in my own research. As far as I can see, other submissions reviewed for scope clearly lack on certain technical criteria. The good thing is that the package is already available for everyone to use. I hope that the process converges on a sound decision, and bring my apologies for expressing a certain amount of surprise.
