### 2025-02-27 #7796 @danielskatz
@Rhyd-Lewis - this has passed the JOSS scope review, and I'll next look for an editor. We're short on editors currently, but are bringing some new ones on board soon, so this may take slightly longer than normal and longer than we would prefer.
