### 2025-03-04 #7871 @editorialbot
[REVIEW]: GCol: A High-Performance Python Library for Graph Colouring
**Submitting author:** @Rhyd-Lewis (<a href="http://orcid.org/0000-0003-1046-811X">Rhyd Lewis</a>)
**Repository:** https://github.com/Rhyd-Lewis/GCol
**Branch with paper.md** (empty if default branch): josspaper
**Version:**  v1.4.2
