### 2025-03-04 #7871 @paulbrodersen
Hi @Rhyd-Lewis, pleasure to make your acquaintance. W.r.t. item 5 on our checklists, could you expand a bit on the contributions of your co-author Geraint Palmer? None of the commits seem to list him as author or co-author (though I am aware that the software was developed offline and uploaded at a late stage in the development cycle).
