### 2025-03-04 #7871 @paulbrodersen
W.r.t Documentation / item 5: Automated tests:

The test suite looks great, the selection of test cases is self-explanatory, and the whole suite passes without skipping any tests in a reasonable amount of time. Two comments though:

1. While anybody likely to execute the test suite will know how to execute a pytest test suite, please include instructions in the documentation how to execute the tests. It allows me to check that box, and it can help with trouble-shooting support requests in the future.
2. pytest should be an optional dependency:

```python
setup(
    name="gcol",
    ...
    extras_require = {
        'testing':  ["pytest"]
    }
)
```
