### 2025-03-04 #7871 @paulbrodersen
Ditto for pandas, which is used in some of the examples:

```python
setup(
    name="gcol",
    ...
    extras_require = {
        'testing':  ["pytest"],
        'documentation' : ["pandas"],
    }
)
```
