### 2025-03-05 #7871 @Rhyd-Lewis
Hi @paulbrodersen -- thank you for your comments and your time.

I have created a new [branch](https://github.com/Rhyd-Lewis/GCol/tree/test_extra_require_in_setup) with the above arguments added to the setup.py file. Once it's been tested, I'll merge it into main.

Regarding your other question, Geraint is my colleague at Cardiff university who has been very helpful in preparing this library. We pair programmed using my github account and he helped to oversee the packaging, paper preperation, linking to gitub, and conversion of my existing c++-based algorithms into python implementations.

Thanks

Rhyd
