### 2025-03-10 #7871 @IvanIsCoding
I have reviewed the software and my recommendation for the editor is to publish. It is on par with other software published in the journal. My opinion is that it would be a welcome addition to JOSS.

## Documentation

For the documentation I followed the [`gcol` documentation published in ReadTheDocs](https://gcol.readthedocs.io/en/latest/) that is linked in the repository. Overall, the software is neatly documented. I will disclaim that I only read the documentation, not tried to generate it.

The introduction section for the package is excellent and it is full of examples. The visualizations make the purpose of the package obvious.

The case study was a pleasant surprise. It covers exam timetabling, which is a a topic I generally see Integer Linear Programming as a more commonly proposed solution. I was curious and found that under the hood the final solution used Tabu search. But the abstraction of modeling the problem as a coloring problem was clever.

There is a performance section in the documentation and I agree that it should be there. I will touch on this point again at the "Paper" section.

Lastly, every algorithm exposed by `gcol` has a very detailed docstring. I found them more than satisfactory. The proper citations in the docstrings are a bonus as well, generally packages are missing those.

## Software

With regards to the software, I believe `gcol` is a significant scholarly contribution. I will flag to the editor that the corresponding author has published a book with related code in C++ as cited in the paper. However, the authors have packaged it as a cohesive piece rewritten in Python, which I agree is more appropriate given the target audience. In addition, they also used the de-facto standard for representing graphs in Python (NetworkX) which further extends the reach of the library.

From a Software Engineering point of view, the library is easy to install and the tests are satisfactory. The authors have properly packaged `gcol` with setuptools and distributed in in PyPI. The authors also chose a popular Python testing library (`pytest`) which made running the tests trivial.

Lastly, I ran [`pytest-cov`](https://pypi.org/project/pytest-cov/) against the codebase. The result was 88% coverage for `gcol`. Code coverage is not everything but the numbers validate the authors' intent to write a complete test suite.

## Paper

This is the only section that I have suggestions for the authors to change.

### Benchmark

My suggestion for the authors is to remove this section from the paper. Do mention throughout the paper that the algorithms implemented by `gcol` are faster and that the heuristics included are better. Do refer readers to the "Performance" section of the documentation.

However, the word count for JOSS is limited and there is not nearly enough space to include a detailed benchmark. There are also caveats about the correctness of the benchmarks, if the authors included all the relevant libraries in the benchmarks, the quality of the coloring found by each library, and more. I got a lot of pushback in the review of [`rustworkx`](https://github.com/openjournals/joss-reviews/issues/3968). I think the `gcol` paper will be vunerable to the same kind of criticism `rustworkx` faced if it keeps the benchmark section.

Moreover, as someone that has benchmarked NetworkX the reality is that performance is not the primary goal of NetworkX. NetworkX has many strenghts like its API and trying to cover as many areas of network science as possible. The NetworkX developers originally did not try
to write the most performant code, it is known to be slow. That is why there are some efforts like `nx-cugraph` to offload computations to the GPU and try to make the library faster.

### Statement of Need: suggestion to add package

Given the audience, I would also include a citation to [IGraph/M](https://joss.theoj.org/papers/10.21105/joss.04899). I am not a user of Mathematica, but I know it has its niche with mathematicians just like SageMath. IGraph/M does implement some coloring related methods, I think igraph uses SAT under the hood to solve the problem. It is not a optimized as `gcol` but I think it fits with the other software cited.

### Citations

These are minor details, but:

* [The authors should cite NetworkX following the guidelines of the package](https://networkx.org/documentation/stable/index.html#citing)
* [The authors should cite SageMath following the guidelines of the package](https://wiki.sagemath.org/Publications_using_SageMath)
* [The authors probably should cite the LEMON paper instead of the link](https://doi.org/10.1016/j.entcs.2011.06.003)
* [The authors should cite ColPack following the guidelines of the package](https://github.com/CSCsw/ColPack?tab=readme-ov-file#the-best-source-for-citing-this-work)
* For Vizing's theorem and DSatur there has to be a better academic source than Wikipedia. You don't have to cite the originals. Feel free to pick any chapter from a mathematics book that cover the topics and cite that instead.

Last but not least, the bot will flag missing DOIs for publications when applicable. So do revisit the citations to include the DOIs.
