### 2025-03-10 #7871 @paulbrodersen
Apart from the performance test for the exact algorithm, I finished running all other performance tests in the documentation (albeit using only 5 seeds instead of 50), and I can confirm the results.

As a comment (i.e. not an essential TODO for the purpose of this review) I would like to add that for the results on equitable coloring, I think a 1-2 sentence explanation why a smaller number of colors increases the running time would be nice.
