### 2025-03-10 #7871 @paulbrodersen
Regarding the performance/benchmarking section, I slightly disagree with @IvanIsCoding . Computing a graph colouring can get expensive pretty quickly, so having some indication of the performance of an implementation seems quite critical for making an informed choice.  Yes, performance comparisons to networkx are a bit of a strawman, but then again, networkx (for better or worse) is the tool that most people will reach for unless they are shown rather explicitly that much better options exist.

@IvanIsCoding does point out correctly that this ~500 word section pushes the paper past 1500 words, which is quite a bit more than [the 1000 word limit](https://joss.readthedocs.io/en/latest/paper.html). However, given that JOSS is an online publication, perhaps an exception can be made. Maybe the editor could weigh in, @vissarion ?
