### 2025-03-10 #7871 @IvanIsCoding
Also, I will add that if this graph from the documentation only compared `gcol` and removed NetworkX it would be a great example of comparing the performance relative to the strategies the authors implemented. It would have its place in the paper.

![asymptotic](https://gcol.readthedocs.io/en/latest/_images/output_3_1.png)

But when we include NetworkX we'd need to ask why it didn't include igraph, why did the authors use a randomized graph instead of a benchmark for graph coloring, and so on.
