### 2025-03-10 #7871 @paulbrodersen
> > [@IvanIsCoding](https://github.com/IvanIsCoding) has covered the references pretty well. Personally, given that GCol is a python package, I would prefer a reference of the python bindings of [igraph](https://python.igraph.org/en/0.11.8/) rather than the Mathematica bindings. [graph-tool](https://graph-tool.skewed.de/) is another popular python graph library that may warrant a 0.5-1 sentence discussion.
>
> That is a fair point about the Mathematica bindings. I tried to find coloring functions in igraph for Python, but I couldn’t find it. [The Mathematica version however had a whole documentation section dedicated to it](http://szhorvat.net/mathematica/IGDocumentation/#graph-colouring).

The documentation for the C functions can be found [here](https://igraph.org/c/doc/igraph-Coloring.html#igraph_is_perfect).

I couldn't find these methods mentioned in the python documentation, but the python bindings do exist in the latest version.

![Image](https://github.com/user-attachments/assets/95c0b187-fe48-4a4a-98b6-cc77e4b34eb1)

That being said, graph coloring appears to be treated more extensively in the Mathematica version, so maybe that is the better point of comparison.
