### 2025-03-14 #7871 @Rhyd-Lewis
Thank you both, @paulbrodersen and @IvanIsCoding, for your very helpful reviews. Geraint and I will make the necessary changes (and give replies) in the coming days. Given the discussions above, we are minded to simply delete the paper section on performance and just make reference to the relevant parts of the the documentation instead. If we have misunderstood, please let us know and we will reconsider.

We'll be back with everything soon....

Thanks again

RL and GP
