+++ paper/paper.bib
@@ -1,149 +1,188 @@
-@Misc{Boost,
-  Title = {Boost Graph Library},
-  year = {2025},
-  HowPublished = {\url{https://www.boost.org/doc/libs/1_46_1/libs/graph/doc/index.html}},
-  Note = {Accessed 2025-01-20},
+@misc{Boost,
+  title        = {The {B}oost Graph Library},
+  year         = 2025,
+  note         = {Accessed 2025-01-20},
+  howpublished = {\url{https://www.boost.org/doc/libs/1_46_1/libs/graph/doc/index.html}}
 }
 
-@misc{colpack,
-  title = {{ColPack}},
-  year = {2025},
-  HowPublished = {\url{https://cscapes.cs.purdue.edu/coloringpage/software.htm}},
-  Note = {Accessed 2025-01-20},  
+@article{Gebremedhin2013,
+  title        = {Col{P}ack: Graph coloring software for derivative computation and beyond},
+  author       = {Gebremedhin, A. and Nguyen, D. and Patwary, M and Pothen, A.},
+  year         = 2013,
+  journal      = {ACM Transactions on Mathematical Software},
+  volume       = 40,
+  number       = 1,
+  pages        = {1--31},
+  doi          = {10.1145/2513109.2513110}
 }
 
-@Book{Cranston2024,
-  Title = {Guide Coloring Methods},
-  Author = {Cranston, D.},
-  Publisher = {Cranston, Richmond, Virginia},
-  Year = {2024},
-  ISBN = {979-8-218-46240-0},
-  url = {https://graphcoloringmethods.com/}
+@article{Brelaz1979,
+  title        = {New methods to color the vertices of a graph},
+  author       = {Brelaz, D.},
+  year         = 1979,
+  journal      = {Communications of the ACM},
+  volume       = 22,
+  number       = 4,
+  pages        = {251--256},
+  doi          = {10.1145/359094.359101}
 }
 
-@Misc{Culberson1994,
-  Title = {Culberson's Node Coloring Implementations},
-  Author = {Culberson, J.},
-  HowPublished = {\url{https://www3.cs.stonybrook.edu/~algorith/implement/culberson/distrib/}},
-  Note = {Accessed 2025-01-20},
-  Year = {1994}
+@inproceedings{Hagberg2008,
+  title        = {Exploring network structure, dynamics, and function using {N}etwork{X}},
+  author       = {Hagberg, A. and Schult, D. and Swart, P.},
+  year         = 2008,
+  month        = {Aug},
+  booktitle    = {Proceedings of the 7th Python in Science Conference (SciPy2008)},
+  address      = {Pasadena, CA},
+  pages        = {11--15},
+  doi          = {10.25080/TCWV9851},
+  editor       = {G. Varoquaux and T. Vaught and J. Millman}
 }
 
-@Misc{DSatur,
-  Title = {{DSatur}},
-  year = {2025},
-  HowPublished = {\url{https://en.wikipedia.org/wiki/DSatur}},
-  Note = {Accessed 2025-01-20},
+@article{Dezso2011,
+  title        = {{LEMON} -- An open source {C}++ graph template library},
+  author       = {Dezso, B. and Juttner, A. and Kovacs, P.},
+  year         = 2011,
+  journal      = {Electronic Notes in Theoretical Computer Science},
+  volume       = 264,
+  number       = 5,
+  pages        = {23--45},
+  doi          = {10.1016/j.entcs.2011.06.003}
 }
 
-@Misc{GColDocs,
-  Title = {{GCol}: A Library for Graph Coloring},
-  year = {2025},
-  HowPublished = {\url{https://gcol.readthedocs.io/en/latest/}},
[…]
