+++ paper/paper.md
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: School of Mathematics, Cardiff University, Wales, United Kingdom
    index: 1
-date: 4 February 2025
+date: 21 March 2025
 bibliography: paper.bib
 ---
 
@@ -41,11 +41,11 @@ routines.\label{fig:demo}](output_00_01_02_combined.png){width="350pt"}
 - An *edge colouring* is an assignment of colours to the edges of a graph so
   that all adjacent edges have different colours. The smallest number of colours
   needed for the edges of a graph $G$ is known at the chromatic index, denoted
-  by $\chi'(G)$. According to Vizing's theorem [@VizingsTheorem], $\chi'(G)$ is
+  by $\chi'(G)$. According to Vizing's theorem, $\chi'(G)$ is
   either $\Delta(G)$ or $\Delta(G)+1$, where $\Delta(G)$ is the maximum degree
-  in $G$. Identifying $\chi'(G)$ is also NP-hard. 
+  in $G$. Identifying $\chi'(G)$ is also NP-hard [@Toft2021]. 
 - A *face colouring* is an assignment of colours to the faces of a planar
-  embedding so that all adjacent faces have different colours. Note, that planar
+  embedding so that all adjacent faces have different colours. Note that planar
   embeddings only exist for planar graphs. The smallest number of colours needed
   to colour the faces of a planar embedding is known as its face chromatic
   number. Due to the Four Colour Theorem, this number never exceeds four and,
@@ -54,18 +54,18 @@ routines.\label{fig:demo}](output_00_01_02_combined.png){width="350pt"}
 
 Graph colouring has applications in many practical areas including timetabling,
 sports league scheduling, designing seating plans, code optimisation, and
-solving Sudoku puzzles [@Lewis2021Book]. It is also a topic of great theoretical
+solving Sudoku puzzles [@Lewis2021Book]. It is also a topic of theoretical
 interest [@Cranston2024] that often appears in university-level courses on graph
 theory, algorithms, and combinatorics.
   
-GCol is a new, open-source Python library for graph colouring built on top of
-the well-known NetworkX library [@networkx]. It provides easy-to-use,
+GCol is a new, open-source Python library for graph colouring that is built on top of
+the well-known NetworkX library [@Hagberg2008]. It provides easy-to-use,
 high-performance algorithms for the above three problems, as well as routines
 for equitable colouring, weighted colouring, pre-colouring, maximum independent
 set identification, and solution visualisation. The following code snippet shows
 how to use the library to create a dodecahedral graph $G$, colour its nodes
 using $\chi(G) = 3$ colours, and then output the solution in textual and
-diagrammatic form (see \autoref{fig:snippet}):
+diagrammatic form (see \autoref{fig:snippet}).
 
 ```python
 >>> import networkx as nx
@@ -88,41 +88,47 @@ Node coloring of G = {0: 0, 1: 1, 19: 1, 10: 1, 2: 0, 3: 2, 8: 0, 9: 2, 18: 0,
 Open-source resources for graph colouring have existed for some time, primarily
 for node colouring. An early example is the PL/I code for node colouring
 included in the 1979 paper of [@Leighton1979]. Stand-alone C-based
-resources were also made available online in the mid-1990s
-[@Culberson1994, @Trick1994]. A Java-based package implementing the methods of
+resources were also made available online in the mid-1990s due to 
+[@Culberson1994] and [@Trick1994]. A Java-based package implementing the methods of
 [@Culberson1994] is now also available [@Shah2020].
   
 Graph colouring functionality is also included in some popular open-source C++
 libraries. [@Boost] uses a simple constructive heuristic
-for node colouring, while [@Lemon] includes a method for
[…]
