### 2025-03-21 #7871 @Rhyd-Lewis
Dear @paulbrodersen, @IvanIsCoding, and @vissarion

Thank you again for taking the time to review our submission and for your positive comments on the GCol library. We have now uploaded the newer versions of the .md and .bib files, taking your comments into account.

Based on the feedback, we have decided to remove the Performance section of the paper (keeping only its first and last sentences, which we have merged into the preceding section). The reviewers had slightly differing opinions on this, but we decided to follow this path to reduce the word count, and because the GCol’s documentation already has a more comprehensive analysis of the algorithms.

Following the recommendations of the reviewers, we have also added citations for Graph-Tool and IGraph, as you will see in the Statement of Need section. As advised, we have noted the existence of the R- and C- based versions of IGraph. We have also mentioned the IGraph/M package, noting that Mathematica is proprietary software.

@paulbrodersen: Thank you for the comment regarding the performance of the equitable colouring algorithms. We will add further details to the documentation.

Finally, we have also changed the references as advised. DOIs have been added for all references where they are available, and citations to packages like NetworkX and LEMON now follow their official guidelines. We have also removed references to Wikipedia and replaced them with relevant articles in the literature.

Thank you again for your support. And please let us know if we can do anything further.

RL and GP
