### 2025-03-22 #7871 @IvanIsCoding
I have read the new version and I confirm it addresses all my earlier comments. I still recommend for the editor is to publish.

On a minor note, I think there is an issue with the SageMath bibtex, the citation is showing as a `?` on the paper. But that is a clerical mistake, I checked the "References" box in the review list.
