### 2025-03-24 b94f642
Update paper.md
Corrected SageMath reference
