### 2025-03-24 #7871 @Rhyd-Lewis
Thanks for the feedback @IvanIsCoding

We have corrected the SageMath reference.

Best regards,

RL and GP
