### 2025-03-26 Rhyd-Lewis/GCol#2 @vissarion
JOSS review
Some minor comments related to JOSS review.

- You can use the current version of boost when citing boost graph i.e. version 1.87 (line 117 in manuscript)
- "Python" is capitalized in the text but with lower-case in the references
- https://doi.org/10.6028/jres.084.024 is a valid doi for Leighton, F. (1979). A graph coloring algorithm for large scheduling problems. Journal of Research of the National Bureau of Standards, 84(6), 489–506
- In some places it seems that an alternative citation format could be used, e.g. line 61
  "...in the algorithm suite of (Lewis, 2021a)..." --> "...in the algorithm suite of Lewis (2021a)..."
 for reference `@author:2001`  ->  "Author et al. (2001)" while  `[@author:2001]` -> "(Author et al., 2001)"
