### 2025-03-26 #7871 @vissarion
@Rhyd-Lewis there is a recommendation for publication from both reviewers!
From my side, I have created an issue with some minor comments https://github.com/Rhyd-Lewis/GCol/issues/2
