### 2025-03-27 93caa81
Update paper.bib
Added DOI and version numbers
