+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @misc{Boost,
-  title        = {The {B}oost Graph Library},
+  title        = {The {B}oost Graph Library, version 1.8.7},
   year         = 2025,
   note         = {Accessed 2025-01-20},
   howpublished = {\url{https://www.boost.org/doc/libs/1_46_1/libs/graph/doc/index.html}}
@@ -32,7 +32,7 @@
   author       = {Hagberg, A. and Schult, D. and Swart, P.},
   year         = 2008,
   month        = {Aug},
-  booktitle    = {Proceedings of the 7th Python in Science Conference (SciPy2008)},
+  booktitle    = {Proceedings of the 7th {P}ython in Science Conference ({SciPy2008})},
   address      = {Pasadena, CA},
   pages        = {11--15},
   doi          = {10.25080/TCWV9851},
@@ -93,7 +93,7 @@
 }
 
 @misc{Goblin,
-  title        = {The {G}oblin graph library},
+  title        = {The {G}oblin graph library, version 2.8},
   year         = 2025,
   note         = {Accessed 2025-01-20},
   howpublished = {\url{https://goblin2.sourceforge.net/}}
@@ -106,7 +106,8 @@
   journal      = {Journal of Research of the National Bureau of Standards},
   volume       = 84,
   number       = 6,
-  pages        = {489--506}
+  pages        = {489--506},
+  doi          = {10.6028/jres.084.024}
 }
 
 @book{Lewis2021Book,
@@ -167,7 +168,7 @@
 }
 
 @article{Peixoto2014,
- title = {The graph-tool python library},
+ title = {The graph-tool {P}ython library},
  url = {http://figshare.com/articles/graph_tool/1164194},
  doi = {10.6084/m9.figshare.1164194},
  journal = {figshare},
