### 2025-03-27 9f33804
Update paper.md
Tidied inline references as advised by editor
