### 2025-04-01 #7871 @Rhyd-Lewis
Hi @vissarion

I have updated the metadata in the zenodo record so that the title and authors match.

https://zenodo.org/records/15111694

I think that is everything, but please let me know if I've missed anything else.

Thanks (and sorry for not getting it right last time)
