### 2025-04-01 #7871 @danielskatz
👋 @Rhyd-Lewis - this looks almost ready to accept. The one issue I see is that in the Shah reference, java should be Java. Can you change this in the bibtex to {J}ava? Once this is done, I would be happy to accept and publish this.
