### 2025-04-02 5dc213f
Update paper.bib
Capital J in "Java"
