+++ paper/paper.bib
@@ -111,7 +111,7 @@
 }
 
 @book{Lewis2021Book,
-  title        = {Guide to Graph Colouring: Algorithms and Applications},
+  title        = {A Guide to Graph Colouring: Algorithms and Applications},
   author       = {Lewis, R.},
   year         = 2021,
   publisher    = {Springer Cham},
@@ -137,7 +137,7 @@
 }
 
 @article{Shah2020,
-  title        = {{JCOL}: A Java package for solving the graph coloring problem},
+  title        = {{JCOL}: A {J}ava package for solving the graph coloring problem},
   author       = {Shah, S.},
   year         = 2020,
   journal      = {Journal of Open Source Software},
